feat: optimizes jwe session codec encode/decode
